requirements.txt
asgiref==3.8.1
cfgv==3.4.0
distlib==0.3.8
Django==5.0.6
django-webpack-loader==3.1.0
filelock==3.15.4
identify==2.5.36
nodeenv==1.9.1
pillow==10.3.0
platformdirs==4.2.2
pre-commit==3.7.1
psycopg2-binary==2.9.9
PyYAML==6.0.1
ruff==0.5.0
setuptools==70.0.0
sqlparse==0.5.0
virtualenv==20.26.3
wheel==0.43.0
